Add deleted images from supabase when image is changed to new one or cabin is deleted

+++ src/services/apiCabins.js
@@ -11,8 +11,15 @@ export async function getCabins() {
   return data;
 }
 
-export async function createEditCabin(newCabin, id) {
+export async function createEditCabin(
+  newCabin,
+  id,
+  isEditSession = false,
+  imageToDelete
+) {
+  const oldImagePath = imageToDelete?.split("/").at(8);
   const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
+  const uploadNewImage = Boolean(hasImagePath);
 
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     "/",
@@ -36,28 +43,39 @@ export async function createEditCabin(newCabin, id) {
 
   if (error) {
     console.error(error);
-    throw new Error("Cabin could not be created");
+    throw new Error("Cabin could not been created");
   }
 
-  //2. Upload image
-  const { error: storageError } = await supabase.storage
-    .from("cabin-images")
-    .upload(imageName, newCabin.image);
+  //2. Upload/Delete image
+  if (!uploadNewImage) {
+    //A) Upload new image
+    const { error: storageError } = await supabase.storage
+      .from("cabin-images")
+      .upload(imageName, newCabin.image);
 
-  //3. Delete the cabin IF there was an error uploading image
+    //3. Delete the cabin IF there was an error uploading image and is not in edit mode
 
-  if (storageError) {
-    await supabase.from("cabins").delete().eq("id", data.id);
-    console.error(storageError);
-    throw new Error(
-      "Cabin image could not be uploaded and the cabin was not created"
-    );
+    if (storageError && !isEditSession) {
+      await supabase.from("cabins").delete().eq("id", data.id);
+      console.error(storageError);
+      throw new Error(
+        "Cabin image could not be uploaded and the cabin was not created"
+      );
+    }
+
+    //B) Delete old image from bucket if image is in edit mode and new image was uploaded
+    if (isEditSession) {
+      await supabase.storage.from("cabin-images").remove(oldImagePath);
+    }
+
+    return data;
   }
-  return data;
 }
 
-export async function deleteCabin(id) {
+export async function deleteCabin(id, image) {
+  const deleteImage = image.split("/").at(8);
   const { data, error } = await supabase.from("cabins").delete().eq("id", id);
+  await supabase.storage.from("cabin-images").remove(deleteImage);
 
   if (error) {
     console.error(error);
